+++ .gitignore
@@ -1,0 +1,1 @@
+**/.DS_Store
